Re-organize histogram code and set DefaultHistogramReader as hist reader for all formats
