+++ pims/api/histograms.py
@@ -27,7 +27,7 @@ from pims.api.utils.models import CollectionSize, HistogramType
 from pims.api.utils.parameter import imagepath_parameter
 from pims.api.utils.response import response_list
 from pims.files.file import HISTOGRAM_STEM, Path
-from pims.files.histogram import argmax_nonzero, argmin_nonzero, build_histogram_file
+from pims.processing.histograms.utils import argmax_nonzero, argmin_nonzero, build_histogram_file
 from pims.utils.iterables import ensure_list
 
 router = APIRouter()
