+++ pims/importer/importer.py
@@ -29,7 +29,7 @@ from pims.files.file import (
     EXTRACTED_DIR, HISTOGRAM_STEM, ORIGINAL_STEM, PROCESSED_DIR, Path,
     SPATIAL_STEM, UPLOAD_DIR_PREFIX
 )
-from pims.files.histogram import Histogram, build_histogram_file
+from pims.files.histogram import Histogram
 from pims.files.image import Image
 from pims.formats import AbstractFormat
 from pims.formats.utils.factories import FormatFactory, SpatialReadableFormatFactory
@@ -37,6 +37,7 @@ from pims.importer.listeners import (
     CytomineListener, ImportEventType, ImportListener,
     StdoutListener
 )
+from pims.processing.histograms.utils import build_histogram_file
 from pims.utils.strings import unique_name_generator
 
 log = logging.getLogger("pims.app")
